+++ ChangeLog.md
@@ -60,6 +60,8 @@ Different versions of mealpy in terms of passing hyper-parameters. So please car
 + Add `Tree Seed Algorithm (TSeedA)` in `bio-based` group
 + Add `Secretary Bird Optimization Algorithm (SBOA)` in `bio-based` group
 + Add `Chameleon Swarm Algorithm (ChameleonSA) and its improved version IChameleonSA` in the `swarm-based` group
++ Add `White Shark Optimizer (WSO)` in the `swarm-based` group
+
 
 ---